requirements.txt
xgboost==3.1.3
biopython==1.85
pydantic==2.10.3
shap==0.45.1
matplotlib==3.8.4
Pillow==10.3.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2